apps/unity: Fix rounded headerbar below titlebar
Some apps have been patched to have Unity titlebar on top of the
common headerbars. In some cases however the rounded top corners
of headerbars are still enabled even with the titlebars on top,
creating a weird effect.

Remove border-radius and box-shadow in this case.

Notes: my tests did not reveal any break of gtk theme, but keep
an eye on this change.